+++ blaze-server/src/main/scala/org/http4s/server/blaze/BlazeServerBuilder.scala
@@ -21,7 +21,6 @@ import org.http4s.blaze.pipeline.stages.SSLStage
 import org.http4s.blazecore.{BlazeBackendBuilder, tickWheelResource}
 import org.http4s.server.ServerRequestKeys
 import org.http4s.server.SSLKeyStoreSupport.StoreInfo
-import org.http4s.util.SSLContextFactory
 import org.log4s.getLogger
 import scala.collection.immutable
 import scala.concurrent.{ExecutionContext, Future}
